pyproject.toml
[project]
name = "olx-parse"
version = "0.1.0"
description = ""
authors = [
    {name = "User",email = "tango5555.li@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "aiohttp (>=3.13.2,<4.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "fastapi[uvicorn] (>=0.120.2,<0.121.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "fake-useragent (>=2.2.0,<3.0.0)",
    "cloudscraper (>=1.2.71,<2.0.0)",
    "environs (>=14.4.0,<15.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "aio-pika (>=9.5.7,<10.0.0)",
    "aiomysql (>=0.3.2,<0.4.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
